feat: restructure CodeEditor component
