+++ web/src/components/features/workspace/CodeEditor/CodeEditor.tsx
@@ -9,9 +9,9 @@ import { Analyzer } from '~/services/analyzer'
 import { TargetType } from '~/services/config'
 import { Connect, newMarkerAction, runFileDispatcher } from '~/store'
 import { type WorkspaceState, dispatchFormatFile, dispatchResetWorkspace, dispatchUpdateFile } from '~/store/workspace'
-import { getTimeNowUsageMarkers, wrapAsyncWithDebounce } from '../utils'
-import { attachCustomCommands } from '../commands'
-import { LANGUAGE_GOLANG, stateToOptions } from '../props'
+import { getTimeNowUsageMarkers, wrapAsyncWithDebounce } from './utils'
+import { attachCustomCommands } from './commands'
+import { LANGUAGE_GOLANG, stateToOptions } from './props'
 import { configureMonacoLoader } from './loader'
 import { registerGoLanguageProvider } from './autocomplete'
